requirements.txt
bintrees==2.2.0
bitarray==1.6.3
boofuzz==0.3.0
cffi==1.14.4
click==7.1.2
cryptography==3.3.1
dnspython==2.1.0
Flask==1.1.2
future==0.18.2
impacket==0.9.22
itsdangerous==1.1.0
ipython==7.22.0
Jinja2==2.11.2
ldap3==2.9
ldapdomaindump==0.9.3
MarkupSafe==1.1.1
netaddr==0.8.0
numpy==1.20.2
pcapy==0.11.4
pyasn1==0.4.8
pycparser==2.20
pycryptodomex==3.9.9
pyOpenSSL==20.0.1
PyYAML==5.4.1
scapy==2.4.5
six==1.15.0
Werkzeug==1.0.1
